% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_capm.R
\name{alpha.capm}
\alias{alpha.capm}
\title{CAPM Alpha}
\usage{
alpha.capm(R1, R2)
}
\arguments{
\item{R1}{Portfolio return  as xts}

\item{R2}{Benchmark Returns}
}
\value{
Returns the alpha of the portfolio
}
\description{
Calculates the portfolio alpha
}
\details{
Alpha is a term used in investing to describe a strategy's ability to beat the market, or it's "edge." Alpha is thus also often referred to as “excess return” or “abnormal rate of return,” which refers to the idea that markets are efficient, and so there is no way to systematically earn returns that exceed the broad market as a whole.
}
\examples{
alpha.capm(funds$ret1, funds$rfr)
}
