% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_capm.R
\name{beta.capm}
\alias{beta.capm}
\title{CAPM Beta}
\usage{
beta.capm(R1, R2)
}
\arguments{
\item{R1}{Returns data of the security}

\item{R2}{Returns data of the benchmark security}
}
\value{
Value of the beta of the security
}
\description{
Returns the Beta of Security using the CAPM Model
}
\details{
Beta is a measure of the volatility–or systematic risk–of a security or portfolio compared to the market as a whole.
}
\examples{
beta.capm(funds$ret1, funds$rfr)
}
