% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funds.R
\docType{data}
\name{funds}
\alias{funds}
\title{Sample Portfolio Return}
\format{
A xts with 901 rows and 2 variables:
\describe{
\item{\code{ret1}}{Portfolio Return}
\item{\code{rfr}}{Benchmark Return (Proxy for risk free rate)}
}
}
\usage{
funds
}
\description{
An xts dataset to calculate the portfolio metrics in the package
}
\details{
This is a sample returns data of the portfolio to understand and use the functions of the package.
}
\keyword{datasets}
