% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markowitz_model.R
\name{markowitz.model}
\alias{markowitz.model}
\title{Markowitz Mean-Variance Model}
\usage{
markowitz.model(R1, R2)
}
\arguments{
\item{R1}{Portfolio Returns}

\item{R2}{Benchmark Returns}
}
\value{
Returns the optimum portfolio weights and their risk and return profile.
}
\description{
Calculates the optimum Portfolio weights
}
\details{
Modern portfolio theory (MPT), or mean-variance analysis, is a mathematical framework for assembling a portfolio of assets such that the expected return is maximized for a given level of risk.
}
\examples{
markowitz.model(funds$ret1, funds$rfr)
}
