% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_ratio.R
\name{ratio.information}
\alias{ratio.information}
\title{Information Ratio}
\usage{
ratio.information(R1, R2)
}
\arguments{
\item{R1}{Returns of the portfolio}

\item{R2}{Returns of the benchmark portfolio}
}
\value{
Calculates the information ratio of the portfolio
}
\description{
Calculates the information ratio of the portfolio
}
\details{
The information ratio (IR) is a measurement of portfolio returns beyond the returns of a benchmark, usually an index, compared to the volatility of those returns.
}
\examples{
ratio.information(funds$ret1, funds$rfr)
}
