% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_sharpe.R
\name{ratio.sharpe}
\alias{ratio.sharpe}
\title{Sharpe Ratio}
\usage{
ratio.sharpe(R1, Rf = 0)
}
\arguments{
\item{R1}{Portfolio Returns}

\item{Rf}{Risk Free Rate of Return, Default: 0}
}
\value{
Calculates the Sharpe Ratio of the portfolio
}
\description{
Calculates the Sharpe Ratio of the Portfolio
}
\details{
The Sharpe ratio was developed by Nobel laureate William F. Sharpe and is used to help investors understand the return of an investment compared to its risk.
}
