% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio_treynor.R
\name{ratio.treynor}
\alias{ratio.treynor}
\title{Treynor Ratio}
\usage{
ratio.treynor(R1, Rf = 0)
}
\arguments{
\item{R1}{Returns of the portfolio}

\item{Rf}{Returns of the benchmark portfolio}
}
\value{
This function can be used to calculate the Treynor ratio of a portfolio.
}
\description{
Calculates the Treynor ratio of a particular portfolio
}
\details{
The Treynor ratio, also known as the reward-to-volatility ratio, is a performance metric for determining how much excess return was generated for each unit of risk taken on by a portfolio.
}
\examples{
ratio.treynor(funds$ret1)
}
