% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returns_annualized.R
\name{returns.cal}
\alias{returns.cal}
\title{Annualized Returns}
\usage{
returns.cal(R1, freq = 252, geometric = TRUE)
}
\arguments{
\item{R1}{Returns dataset as xts}

\item{freq}{The periodicity of the dataset, Default: 252}

\item{geometric}{Boolean to control the geometric returns and mean annualized returns, Default: TRUE}
}
\value{
Gives annualized returns of data
}
\description{
Returns the annualized returns of a data returns data
}
\details{
An annualized total return is the geometric average amount of money earned by an investment each year over a given time period.
}
\examples{
returns.cal(funds$ret1)
}
