% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_premium.R
\name{risk.premium}
\alias{risk.premium}
\title{Risk Premium of a Security}
\usage{
risk.premium(R1, Rf)
}
\arguments{
\item{R1}{The returns of the security as xts}

\item{Rf}{The risk free rate of return as xts}
}
\value{
Returns the risk premium of the security
}
\description{
This function is used to calculate the risk premium of excess return over the risk free rate
}
\details{
A risk premium is the return in excess of the risk-free rate of return an investment is expected to yield
}
\examples{
risk.premium(funds$ret1, funds$rfr)
}
