% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_deviation.R
\name{semi.deviation}
\alias{semi.deviation}
\title{Semi Deviation/ Down side Deviation}
\usage{
semi.deviation(R1)
}
\arguments{
\item{R1}{Returns dataset of the portfolio}
}
\value{
Calculates the semi deviation of the xts object
}
\description{
Calculates the semi deviation of the xts object
}
\details{
Semi-deviation is a method of measuring the below-mean fluctuations in the returns on investment.
}
\examples{
 semi.deviation(funds$ret1)
}
