% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_error.R
\name{tracking.error}
\alias{tracking.error}
\title{Tracking Error}
\usage{
tracking.error(R1, R2)
}
\arguments{
\item{R1}{Returns of the portfolio}

\item{R2}{Returns of the benchmark}
}
\value{
Calculates the Tracking error of the security
}
\description{
Calculates the Tracking Error
}
\details{
Tracking error is the divergence between the price behavior of a position or a portfolio and the price behavior of a benchmark.
}
\examples{
tracking.error(funds$ret1, funds$rfr)
}
